ログイン入力分のコンポーネント化
