+++ src/components/organisms/index.ts
@@ -1,3 +1,4 @@
 export * from './AppBar'
 export * from './Footer'
+export * from './Login'
 export * from './Logo'
